talentlens-ai/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
pypdf==5.1.0
python-docx==1.1.2
langchain==0.3.13
langgraph==0.2.60
openai==1.58.1
google-generativeai==0.8.3
httpx==0.28.1
